organise the components in a proper manner, added the login popup and made the changes in design
