+++ public/index.html
@@ -6,7 +6,7 @@
     <link rel="icon" href="assets/logo.svg" />    
     <title>Ticaret</title>
   </head>
-  <body id="root">
-    
+  <body>
+    <div id="root"></div>
   </body>
 </html>
